assessment-analysis-project/requirements.txt
requests>=2.25.1
pandas>=1.3.0
python-dotenv>=0.19.0
weasyprint>=54.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0 